Hyperparameter Tuning on Test Set
I changed our hyper parameterization to the test set instead of the training set. We got a slightly lower ROC-AUC than the initial model, but remember the initial model was used on the train set. I think the optimized model is less prone to overfitting, so I added a different interpretation at the end that explains it in details.